+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update SettingsDialog tests after Remote indicator relocation
 ---
 
 ## [0.29.0] - 2026-03-30